re-organize ReviewModal and ReviewActions
